Group common functions as engine_utils
